moved app into bin
