+++ Cargo.toml
@@ -1,3 +1,6 @@
+[workspace]
+members = ["crates/*", "documentation"]
+
 [package]
 name = "excalibur"
 version = "0.2.0"
@@ -9,8 +12,6 @@ authors = ["Primitive Bits, Inc."]
 name = "excalibur"
 path = "bin/src/main.rs"
 
-[workspace]
-members = ["crates/*", "documentation", "bin"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -48,9 +49,7 @@ alloy-rlp = "0.3"
 alloy-pubsub = { git = "https://github.com/alloy-rs/alloy.git", rev = "0ba6b61" }
 alloy-signer = { git = "https://github.com/alloy-rs/alloy.git", rev = "0ba6b61" }
 
-
 reqwest = "=0.11.23"
-
 
 ## Config and parsing
 serde = "=1.0.192"
@@ -93,9 +92,10 @@ criterion = "=0.5.1"
 chrono = { version = "=0.4.31", features = ["serde"] }
 
 [dependencies]
-# main excalibur application
-app = { path = "./crates/app", version = "0.1.0" }
-sim = { path = "./crates/sim", version = "0.1.0" }
+datatypes.workspace = true
+clients.workspace = true
+cfmm_math.workspace = true
+sim.workspace = true
 
 clap = { version = "=4.4.14", features = ["derive"] }
 
@@ -110,6 +110,68 @@ tracing-subscriber.workspace = true
 dotenv.workspace = true
 RustQuant.workspace = true
 statrs.workspace = true
+
+# crates
+
+
+# deps
+iced = { version = "=0.10.0", features = [
+    "system",
+    "tokio",
+    "lazy",
+    "image",
+    "palette",
+    "advanced",
+    "canvas",
+    "svg",
+] }
+iced_aw = { version = "=0.7.0", default-features = false, features = [
+    "menu",
+    "icon_text",
+    "icons",
+    "spinner",
+    "floating_element",
+] }
+plotters-backend = "=0.3.5"
+plotters-iced = "=0.9.0"
+iced-loading-indicator = { git = "https://github.com/BB-301/iced-loading-indicator.git", rev = "a09296f" }
+url = "=2.5.0"
+
+# workspace crates
+
+# revm
+revm.workspace = true
+revm-primitives.workspace = true
+
+# alloy
+alloy-primitives.workspace = true
+alloy-sol-types.workspace = true
+alloy-providers.workspace = true
+alloy-networks.workspace = true
+alloy-rpc-client.workspace = true
+alloy-rpc-types.workspace = true
+alloy-transport-ws.workspace = true
+alloy-rlp.workspace = true
+alloy-pubsub.workspace = true
+alloy-transport-http.workspace = true
+alloy-signer.workspace = true
+reqwest.workspace = true
+
+# workspace deps
+chrono.workspace = true
+arbiter-bindings.workspace = true
+bindings.workspace = true
+async-trait.workspace = true
+reikna.workspace = true
+arbiter-core.workspace = true
+ethers.workspace = true
+directories-next.workspace = true
+plotters.workspace = true
+tokio-util.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+uuid.workspace = true
+
 
 [dev-dependencies]
 # For book
