+++ bin/components/chart/mod.rs
