+++ bin/components/tables/cells/toggle.rs
