+++ bin/controller/portfolio/monolithic/inventory.rs
