+++ bin/controller/portfolio/monolithic/view.rs
