+++ bin/middleware.rs
