+++ bin/model/contacts/list.rs
