+++ bin/model/mod.rs
