+++ bin/model/portfolio.rs
