+++ bin/model/rpcs/mod.rs
