+++ bin/model/user.rs
