+++ bin/tracer.rs
