+++ bin/view/mod.rs
