+++ bin/view/portfolio_view.rs
