+++ bin/view/sidebar.rs
